feature-#7 [계량도구][UI] 계량도구추가 인터페이스 콜백(?!)
- 프래그먼트 EditText 누르면 Activity 이미지 사라짐
