+++ app/src/main/java/cookcook/nexters/com/amoogye/views/tools/tools_list/ToolsViewPageAdapter.kt
@@ -1,4 +1,4 @@
-package cookcook.nexters.com.amoogye.views.tools
+package cookcook.nexters.com.amoogye.views.tools.tools_list
 
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentManager
@@ -7,8 +7,12 @@ import androidx.fragment.app.FragmentPagerAdapter
 class ToolsViewPageAdapter(fm:FragmentManager) : FragmentPagerAdapter(fm) {
     override fun getItem(position: Int): Fragment {
         return when(position) {
-            0 -> { ToolsFragmentNormal() }
-            else -> { ToolsFragmentLife() }
+            0 -> {
+                ToolsFragmentNormal()
+            }
+            else -> {
+                ToolsFragmentLife()
+            }
         }
     }
